+++ hjem.html
@@ -3,6 +3,29 @@
         <link rel="stylesheet" type="text/css" href="static/css/style.css">
     </head>
 
+    <script>
+        const buttonClicked=() => {
+            let felt=document.getElementById("felt").value;
+            let radNummer=document.getElementById("rad").value;
+            let seteNummer=document.getElementById("sete").value;
+            
+            if(felt==="" || radNummer==="" || seteNummer==="") {
+                document.getElementById("ErrorMessage").innerText="Vennligst fyll inn alle feltene!";
+            }
+            else if (isNaN(felt) || isNaN(radNummer) || isNaN(seteNummer)) {
+                document.getElementById("ErrorMessage").innerText="Input må være tall!";
+            }
+            else {
+                document.getElementById("ErrorMessage").innerText="";
+                localStorage.setItem("felt", felt);
+                localStorage.setItem("radNummer", radNummer);
+                localStorage.setItem("seteNummer", seteNummer);
+                // console.log(felt, radNummer, seteNummer); FUNKER
+                window.location.href="/farge"; // Redirekt til websocket.html
+            }
+        }
+    </script>
+
     <body>
         <font size = "5">
         <h1>Hei og velkommen til MGP sitt digitale lysshow!</h1>
@@ -17,9 +40,10 @@
         <input id="sete">
         <br>
 
-        <button>Send!</button>
+        <button onclick="buttonClicked()">Send!</button>
+        <div id="ErrorMessage"></div>
         <br>
 
-        <!-- <img src="static/images/MGPlogo.jpg" /> -->
+        <img src="static/images/MGPlogo.jpg" />
     </body>
 </html>
