+++ websocket.html
@@ -26,10 +26,16 @@
       console.log('Disconnected from server');
     });
 
+    var felt=localStorage.getItem("felt");
+    var radNummer=localStorage.getItem("radNummer");
+    var seteNummer=localStorage.getItem("seteNummer");
+
     setInterval(() => {
         let color = document.querySelector('#color').value;
         socket.emit('update:color', color);
     }, 1000)
+
+    
 
   </script>
 </html>
